Rail_Madad-main/requirements.txt
python-dotenv
google-generativeai
Flask
flask-bcrypt
pymongo
Pillow